swaped the program and stack region of memory and introduced the canary
stack and program were swapped because in the program there might be
absolute calls, so the addresses needed to be ensured to be correct
because the stack size can change at runtime.

the canary has been introduced because it costs pratically nothing and
is a very simple feature that is able to detect stack overflows.